services/identity-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
pydantic-settings==2.7.1
email-validator==2.2.0
argon2-cffi==23.1.0
PyJWT==2.10.1
